fixup! Add simple integration test
